+++ test/inherit_url_test.dart
@@ -23,7 +23,7 @@ class _Init implements RootScreenBase {
   final List<(Enum, Object?)> chain;
 }
 
-NavGraph _mk() => NavGraph(
+ScreenGraph _mk() => ScreenGraph(
       {
         U.home({
           U.item({U.editItem.inherit(U.item)})