+++ test/nav_mirror_url_test.dart
@@ -20,7 +20,7 @@ class _Init implements RootScreenBase {
   final List<(Enum, Object?)> chain;
 }
 
-NavGraph _mk() => NavGraph(
+ScreenGraph _mk() => ScreenGraph(
       {
         U.home({U.account()})
       },